Equipment and Inventory moved to actor
Equipment and Inventory containers are no longer character specific, as the no longer implicitly initialise their event listeners. Rather, an actor now owns the equipment, and the character must explicitly define an EquipmentEventListener for themselves if they want to receive event notifications for it.

This is so we can have a clean and uniform API for calculating stats based on equipment and buffs.