Use the player's real inventory in a trade session
Rather than transferring an item to the trade container from a copy straight away, we now just store a reference to an item that has been offered to be traded. These offers are transferred to the respective containers when the trade is finalised.

When an item is moved around in the inventory, if it was offered up for trade, it is now removed.